requirements.txt
Flask==2.1.2
PyYAML==5.4.1
openpyxl==3.0.10
pandas==1.4.2